Renames getMetaDescriptionKeyword to calculateMetaDescriptionKeyword
